code refactor implemented
