+++ src/utils/appData.js
@@ -1,0 +1,2 @@
+export const difficulties = ["Easy", "Normal", "Hard"];
+export const categories = ["Stuff", "Family", "Health", "Learning", "Leisure", "Work"];
